Fix #83

+++ bin/prepare.sh
@@ -85,7 +85,12 @@ then
 	curl -s -L https://nvidia.github.io/nvidia-docker/$distribution/nvidia-docker.list | sudo tee /etc/apt/sources.list.d/nvidia-docker.list
 
 	sudo apt-get update && sudo apt-get install -y --no-install-recommends nvidia-docker2 nvidia-container-toolkit nvidia-container-runtime
-	cat /etc/docker/daemon.json | jq 'del(."default-runtime") + {"default-runtime": "nvidia"}' | sudo tee /etc/docker/daemon.json
+    if [ -f "/etc/docker/daemon.json" ];
+    then
+        cat /etc/docker/daemon.json | jq 'del(."default-runtime") + {"default-runtime": "nvidia"}' | sudo tee /etc/docker/daemon.json
+    else
+        sudo cp $DIR/../defaults/docker-daemon.json /etc/docker/daemon.json
+    fi
 fi
 sudo systemctl enable docker
 sudo systemctl restart docker
